pyproject.toml
[tool.poetry]
name = "peotry-pro"
version = "0.1.0"
description = ""
authors = ["amelia <tndus2623@gmail.com>"]
readme = "README.md"
packages = [{include = "peotry_pro"}]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.97.0"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.16"
pendulum = "^2.1.2"
loguru = "^0.7.0"
pydentic = "^0.0.1.dev3"
pydantic = "^1.10.9"
google-cloud-bigquery = "^3.11.2"
google-cloud-bigquery-storage = "^2.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
